'Community' endpoints validation.

+++ Persistence/UserCommunityRepository.cs
@@ -1,5 +1,6 @@
 using Core.InterfaceContracts;
 using Core.Models;
+using Infrastructure.Exceptions;
 using Microsoft.EntityFrameworkCore;
 
 namespace Persistence;
@@ -15,6 +16,9 @@ public class UserCommunityRepository: IUserCommunityRepository
 
     public async Task<UserCommunity> CreateAsync(UserCommunity userCommunity)
     {
+        if (!await _context.Communities.AnyAsync(c=>c.Id == userCommunity.CommunityId))
+            throw new KeyNotFoundException("Community id=" + userCommunity.CommunityId + " not found in database.");
+        
         _context.UserCommunity.Add(userCommunity);
         await _context.SaveChangesAsync();
         return userCommunity;
@@ -35,15 +39,17 @@ public class UserCommunityRepository: IUserCommunityRepository
 
     public async Task DeleteByIds(Guid communityId, Guid userId)
     {
-        var userCommunityToDelete = await _context.UserCommunity.FirstAsync(uc =>
+        if (!await _context.Communities.AnyAsync(c=>c.Id == communityId))
+            throw new KeyNotFoundException("Community id=" + communityId + " not found in database.");
+        
+        var userCommunityToDelete = await _context.UserCommunity.FirstOrDefaultAsync(uc =>
             uc.UserId == userId && uc.CommunityId == communityId);
+        
+        if (userCommunityToDelete == null)
+            throw new KeyNotFoundException("User id=" + userId + " does not belong to community id=" + communityId + ".");
+        
         _context.UserCommunity.Remove(userCommunityToDelete);
         await _context.SaveChangesAsync();
-    }
-
-    public Task<List<User>> GetCommunitySubscribersAsync(Guid communityId)
-    {
-        throw new NotImplementedException();
     }
 
     public async Task<int> GetSubscriberCountById(Guid communityId)
@@ -54,15 +60,41 @@ public class UserCommunityRepository: IUserCommunityRepository
 
     public async Task<CommunityRole> GetHighestRoleOfUserInCommunity(Guid communityId, Guid userId)
     {
-        try
-        {
-            var userCommunity =
-                await _context.UserCommunity.FirstAsync(uc => uc.UserId == userId && uc.CommunityId == communityId);
-            return userCommunity.HighestRole;
-        }
-        catch
-        {
-            throw new NullReferenceException("UserCommunity not found");
-        }
+        if (!await _context.Communities.AnyAsync(x => x.Id == communityId))
+            throw new KeyNotFoundException("Community id=" + communityId + " not found in database");
+        
+        var userCommunity =
+            await _context.UserCommunity.FirstOrDefaultAsync(uc => uc.UserId == userId && uc.CommunityId == communityId);
+            
+        if (userCommunity == null)
+            throw new KeyNotFoundException("User id=" + userId.ToString() + "does not belong to community id=" + communityId.ToString());
+            
+        return userCommunity.HighestRole;
+    }
+
+    public async Task ConfirmUserBelongsToClosedCommunity(Guid communityId, Guid userId)
+    {
+        var community = await _context.Communities.FirstOrDefaultAsync(c => c.Id == communityId);
+        
+        if (community == null)
+            throw new KeyNotFoundException("Community id=" + communityId.ToString() + " not found in database.");
+        
+        if (! await _context.UserCommunity.AnyAsync(uc => uc.UserId == userId && uc.CommunityId == communityId) && 
+            community.IsClosed)
+            throw new ForbiddenException("User id=" + userId.ToString() + "does not belong to closed community id="+ communityId.ToString());
+    }
+
+    public async Task<bool> IsUserInCommunity(Guid communityId, Guid userId)
+    {
+        
+        var community = await _context.Communities.FirstOrDefaultAsync(c => c.Id == communityId);
+        
+        if (community == null)
+            throw new KeyNotFoundException("Community id=" + communityId.ToString() + " not found in database.");
+
+        if (!await _context.UserCommunity.AnyAsync(uc => uc.UserId == userId && uc.CommunityId == communityId))
+            return false;
+        
+        return true;
     }
 }
